+++ libraries/SWSPI/CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.1] - 2024-11-01
+- fix #2, data in
+- fix transfer16() byte order
+- add transfer24() and transfer32()
+- add examples
+- minor edits
+
 ## [0.1.0] - 2024-10-29
 - initial version, experimental!